Create memToReg mov tests for all registers width and adapt test script to them. - 16- and 8-bit mov test are disabled, see: #1.

+++ test/memToReg_32bit_heap.cpp
@@ -1,11 +1,11 @@
 #include <stdio.h>
 
 int main(int argc, char** argv) {
-    long long *i = new long long;
+    int *i = new int;
     if (*i){
         printf("xx\n");
     }
     return 0;
 }
 
-// EXPECTED: memToRegShadowCopy. Shadow of reg 0 is: 0xffffffffffffffff.
+// EXPECTED: memToRegShadowCopy. Shadow of reg 0 is: 0xffffffff.
